Hello_Agent/Runner_Level/pyproject.toml
[project]
name = "runner-level"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
